fixed dataBridge
